provision database using Ansible after the ebs volume attachment
Provisioning by `local-exec` does so by running Ansible on the local
machine. As such, Ansible version and the version of community
plugins, specifically the Docker plugin, are dependent on the machine
from which Terraform was run.

These changes also migrate the database IP address from that of the
instance to the Elastic IP, to prevent changes over state updates.