build and deploy all Lambda functions in batch under Terraform
Uses a terraform_data resource with a local-exec provisioner to build
the artifacts using `make build`. The files are uploaded as
`aws_s3_object` resources. Rebuilds are performed on the whole batch
if any of the function source code or Go dependency files change.

Also removes the `deploy-lambas` Makefile target.